* fix(1234-Windows-port-new): fix windows port.

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix FTP passive port

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix dynamodb container creation

* fix(1234-Windows-port-new): fix application startup

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix application startup

* fix(1234-Windows-port-new): fix s3 handler stol issue

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix s3 handler stol issue

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1233-Windows-port-new

* fix(1234-Windows-port-new): add application watchdog

* fix(1234-Windows-port-new): add application watchdog

* fix(1234-Windows-port-new): add application watchdog

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add application watchdog

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): new WindowsService.cpp

* fix(1234-Windows-port-new): new WindowsService.cpp

* fix(1234-Windows-port-new): new WindowsService.cpp

* fix(1234-Windows-port-new): add application log server port

* Merge remote-tracking branch 'origin/main' into fix/1234-Windows-port…

* fix(1234-Windows-port-new): add application log server port

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add application log server port

* fix(1234-Windows-port-new): add application log server port

* fix(1234-Windows-port-new): add application log server port

+++ src/core/include/awsmock/core/JsonUtils.h
@@ -44,7 +44,7 @@ namespace AwsMock::Core::Json {
 
     inline long GetLongValue(const boost::json::value &value, const std::string &name, const long defaultValue = 0) {
         if (AttributeExists(value, name)) {
-            return value.at(name).as_int64();
+            return static_cast<long>(value.at(name).as_int64());
         }
         return defaultValue;
     }
